moving the writeWeights() into the base classes for better debugging and fixing small bug in dropout calculation
